+++ analysis_shape/utils.py
